+++ javascript/modules/common/src/encoding.ts
@@ -1,16 +1,23 @@
 /** Decodes a hex string and returns it as a buffer */
-export function hexStringToBuffer(hexString: string): Buffer {
+export function hexToBytes(hexString: string): Uint8Array {
   if (!/^(0x)?[0-9a-fA-F]+$/.test(hexString)) {
     throw new Error("Invalid hex string");
   } else if (hexString.length % 2 !== 0) {
-    throw new Error("The hex string contains an odd length");
+    throw new Error("The hex string has an odd length");
   }
 
-  return Buffer.from(strip0x(hexString), "hex");
+  hexString = strip0x(hexString);
+  const bytes = [];
+  for (let i = 0; i < hexString.length; i += 2) {
+    bytes.push(
+      parseInt(hexString.substring(i, i + 2), 16),
+    );
+  }
+  return Uint8Array.from(bytes);
 }
 
 /** Encodes a buffer into a hex string */
-export function uintArrayToHex(buff: Uint8Array, prepend0x?: boolean): string {
+export function bytesToHex(buff: Uint8Array, prepend0x?: boolean): string {
   const bytes: string[] = [];
   for (let i = 0; i < buff.length; i++) {
     if (buff[i] >= 16) bytes.push(buff[i].toString(16));
@@ -21,14 +28,14 @@ export function uintArrayToHex(buff: Uint8Array, prepend0x?: boolean): string {
 }
 
 /** Encodes the given big integer as a 32 byte big endian buffer */
-export function bigIntToBuffer(number: bigint): Buffer {
+export function bigIntToBuffer(number: bigint): Uint8Array {
   let hexNumber = number.toString(16);
   while (hexNumber.length < 64) hexNumber = "0" + hexNumber;
-  return Buffer.from(hexNumber, "hex");
+  return hexToBytes(hexNumber);
 }
 
 /** Encodes the given big integer as a 32 byte little endian buffer */
-export function bigIntToLeBuffer(number: bigint): Buffer {
+export function bigIntToLeBuffer(number: bigint): Uint8Array {
   return bigIntToBuffer(number).reverse();
 }
 
@@ -53,10 +60,43 @@ export function strip0x(value: string): string {
   return value.startsWith("0x") ? value.substring(2) : value;
 }
 
-export function hexToBytes(hex: string): Uint8Array {
-  const hexMatch = hex.match(/.{1,2}/g)
-  if (!hexMatch) {
-    throw new Error("invalid hex string")
+/**
+ * Encodes a 0-1 ratio within the given digit precision for storage on a smart contract
+ *
+ * @export
+ * @param {number} ratio
+ * @param {number} digits
+ * @return {*}  {bigint}
+ */
+export function encodeRatio(ratio: number, digits: number): number {
+  if (ratio < 0 || ratio > 1) {
+    throw new Error("The ratio value should range between 0 and 1");
+  } else if (!Number.isInteger(digits) || digits < 1 || digits > 15) {
+    throw new Error(
+      "The number of digits should range between 1 and 15",
+    );
   }
-  return Uint8Array.from(hexMatch.map((byte) => parseInt(byte, 16)));
+  return Math.round(ratio * (10 ** digits));
+}
+/**
+ * Decodes a value received from a smart contract to a number with
+ *
+ * @export
+ * @param {bigint} onChainValue
+ * @param {number} digits
+ * @return {*}  {number}
+ */
+export function decodeRatio(
+  onChainValue: bigint | number,
+  digits: number,
+): number {
+  if (!Number.isInteger(digits) || digits < 1 || digits > 15) {
+    throw new Error(
+      "The number of digits should be a positive integer between 1 and 15",
+    );
+  } else if (onChainValue > 10 ** digits) {
+    throw new Error("The value is out of range");
+  }
+
+  return Number(onChainValue) / (10 ** digits);
 }