+++ javascript/modules/common/test/unit/encoding.test.ts
@@ -3,11 +3,12 @@ import {
   bigIntToLeBuffer,
   bufferLeToBigInt,
   bufferToBigInt,
+  bytesToHex,
+  decodeRatio,
+  encodeRatio,
   ensure0x,
-  hexStringToBuffer,
   hexToBytes,
   strip0x,
-  uintArrayToHex,
 } from "../../src";
 
 describe("Value encoding", () => {
@@ -32,7 +33,7 @@ describe("Value encoding", () => {
     ];
 
     for (let input of inputs) {
-      const result = hexStringToBuffer(input.hex);
+      const result = hexToBytes(input.hex);
       expect(result.join(",")).toEqual(input.serializedBuffer);
     }
   });
@@ -98,7 +99,7 @@ describe("Value encoding", () => {
     ];
 
     for (let item of items) {
-      const hex = uintArrayToHex(item.buffer, item.with0x);
+      const hex = bytesToHex(item.buffer, item.with0x);
       expect(hex).toEqual(item.output);
     }
   });
@@ -161,10 +162,12 @@ describe("Value encoding", () => {
 
     for (let input of inputs) {
       const result = bigIntToBuffer(input.bigint);
-      expect(result.toString("hex")).toEqual(input.hexBuffer);
+      expect(Buffer.from(result).toString("hex")).toEqual(input.hexBuffer);
 
       const leResult = bigIntToLeBuffer(input.bigint);
-      expect(leResult.reverse().toString("hex")).toEqual(input.hexBuffer);
+      expect(Buffer.from(leResult).reverse().toString("hex")).toEqual(
+        input.hexBuffer,
+      );
     }
   });
 
@@ -273,16 +276,69 @@ describe("Value encoding", () => {
     }
   });
 
-  it("Should convert a hex string to a decimal Uint8Array", () => {
+  it("Should return an integer encoded from a float between 1 and 0 and a positive integer number of digits", () => {
+    expect(() => encodeRatio(-0.5, 4)).toThrow(
+      "The ratio value should range between 0 and 1",
+    );
+    expect(() => encodeRatio(5, 4)).toThrow(
+      "The ratio value should range between 0 and 1",
+    );
+    expect(() => encodeRatio(0.5, -1)).toThrow(
+      "The number of digits should range between 1 and 15",
+    );
+    expect(() => encodeRatio(0.5, 18)).toThrow(
+      "The number of digits should range between 1 and 15",
+    );
+
     const inputs = [
-      { in: "61", out: new Uint8Array([97]) },
-      { in: "6161", out: new Uint8Array([97, 97]) },
-      { in: "616162", out: new Uint8Array([97, 97, 98]) },
-      { in: "68656c6c6f207468657265", out: new Uint8Array([104, 101, 108, 108, 111, 32, 116, 104, 101, 114, 101]) }
-    ]
+      { in: [0.5, 1], out: 5 },
+      { in: [0.5, 4], out: 5000 },
+      { in: [0.5, 10], out: 5000000000 },
+      { in: [0.25, 1], out: 3 },
+      { in: [0.25, 2], out: 25 },
+      { in: [0.251, 2], out: 25 },
+      { in: [0.251, 3], out: 251 },
+      { in: [0.25, 4], out: 2500 },
+      { in: [0.25, 10], out: 2500000000 },
+    ];
+
     for (let input of inputs) {
-      const result = hexToBytes(input.in);
-      expect(result.toString()).toEqual(input.out.toString());
+      const result = encodeRatio(input.in[0], input.in[1]);
+      expect(result).toEqual(input.out);
     }
+  });
+
+  it("Should decode a float from a given bigint and a number of digits", () => {
+    expect(() => encodeRatio(0.5, -1)).toThrow(
+      "The number of digits should range between 1 and 15",
+    );
+    expect(() => encodeRatio(0.5, 18)).toThrow(
+      "The number of digits should range between 1 and 15",
+    );
+
+    const inputs = [
+      { bigint: BigInt(5), digits: 1, out: 0.5 },
+      { bigint: BigInt(5456), digits: 4, out: 0.5456 },
+      { bigint: 5, digits: 1, out: 0.5 },
+      { bigint: 5456, digits: 4, out: 0.5456 },
+      { bigint: BigInt("1"), digits: 9, out: 0.000000001 },
+      { bigint: BigInt("367483947"), digits: 9, out: 0.367483947 },
+      { bigint: 1, digits: 9, out: 0.000000001 },
+      { bigint: 367483947, digits: 9, out: 0.367483947 },
+    ];
+
+    for (let input of inputs) {
+      const result = decodeRatio(input.bigint, input.digits);
+      expect(result).toEqual(input.out);
+    }
+
+    expect(() => decodeRatio(Number.MAX_SAFE_INTEGER + 1, 12)).toThrow(
+      "The value is out of range",
+    );
+    expect(() => decodeRatio(BigInt("512345898367483947"), 12)).toThrow(
+      "The value is out of range",
+    );
+    expect(() => decodeRatio(10 ** 2, 1)).toThrow("The value is out of range");
+    expect(() => decodeRatio(10 ** 10, 9)).toThrow("The value is out of range");
   });
 });